Update circleDetection.gsql
Cleaning up previously visited list accumulators greatly improves memory usage for this algorithm. This is done by removing the edgeTupleList from the source in the post-Accum step as the new paths are already forwarded to the next vertexes in the path. To my understanding, at that point, there is no need to store the paths in source vertex.